+++ app/controllers/landing_page_controller.rb
@@ -1,13 +1,11 @@
 class LandingPageController < ApplicationController
-  include ChallengeFilter
-
   before_action :terminate_challenge, only: [:index]
   before_action :set_filters, only: [:index]
 
   respond_to :html, :js
 
   def index
-    @challenges = filter_challenge
+    @challenges = ChallengeFilterService.new(params).filter_challenge
     @challenges = @challenges
                       .includes(:organizers)
                       .where(private_challenge: false)
@@ -40,5 +38,20 @@ class LandingPageController < ApplicationController
   def host
     @page_title     = "Organize"
     @challenge_call = ChallengeCall.first
+  end
+
+  private
+
+  def set_filters
+    @categories = Category.all
+    @status     = challenge_status
+    @prize_hash = { prize_cash:     'Cash prizes',
+                    prize_travel:   'Travel grants',
+                    prize_academic: 'Academic papers',
+                    prize_misc:     'Misc prizes' }
+  end
+
+  def challenge_status
+    params[:controller].eql?("landing_page") ? Challenge.statuses.keys - ['draft'] : Challenge.statuses.keys
   end
 end
